testicles

+++ commands/done.js
@@ -1,4 +1,5 @@
 const { Command } = require('smooth-discord.js');
+const { RichEmbed } = require('discord.js');
 const { stripIndents } = require('common-tags');
 const { error } = require('../util.js');
 const requests = require('../models/request.js');
@@ -25,7 +26,18 @@ module.exports = class DoneCommand extends Command {
 		`);
 
 		const logMessage = await message.guild.channels.get('351269513737666560').fetchMessage(request.messageId);
-		await logMessage.delete().catch(console.log);
+
+		if (!logMessage) return message.reply('No request message was found. It must\'ve been deleted!');
+
+		const embed = new RichEmbed()
+			.setDescription(stripIndents`
+				**User**: ${message.author.username}
+				**Request**: ${args}
+				**Status**: Complete
+			`)
+			.setColor(0x4fdd24);
+
+		await logMessage.edit({ embed }).catch(console.log);
 		await message.delete();
 		await message.reply(`Completed task: ${request.id}`);
 	}
